Upset.jl Documentation

Upset.jl Documentation

Functions

Index

Functions

Upset.plot_upset — Method

plotupset(dictin)

Compute an upset plot giving a dictionary of lists. Each list in the dictionary is a list of the unique elements that are present in a set. Used for comparing presence of unique items across groups.

Requires that there exists more than one list in the dictionary in order to look for overlap.

Also requires that at least one list has one item present in it.

Examples

julia> a =    input_dict = Dict(
               "list1" => ['l','t','A','B'],
               "list2" => ['l','t','A','B','C','l'])

Dict{String, Vector{Char}} with 2 entries:
  "list1" => ['l', 't', 'A', 'B']
  "list2" => ['l', 't', 'A', 'B', 'C', 'l']

julia> result = plot_upset(a) 

2-element Vector{Any}:
 Plot{Plots.GRBackend() n=8}
Captured extra kwargs:
  Series{3}:
    series_annotations_align: right
  Series{4}:
    series_annotations_align: right

 Vector{Any}[[4, 1], ["list1", "list2", "list2"]]

Index

Upset.plot_upset